third_party/WebKit/LayoutTests/external/wpt/resource-timing/single-entry-per-resource.html

<!DOCTYPE HTML>
<meta charset=utf-8>
<meta name="timeout" content="long">
<title>One resource when reusing data</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<body>
<script>
async_test((t) => {
  let img_entries = 0;
  const img_url = "resources/blue.png";
  const observer = new PerformanceObserver(
    function (entryList, obs) {
      const resourceEntries = entryList.getEntriesByType("resource");
      if (resourceEntries.length > 0) {
        for (let resourceEntry of resourceEntries) {
          if (resourceEntry.name.indexOf(img_url) != -1) {
            ++img_entries;
          }
        }
        // When this mark is received, all resource timing entries for the
        // images will have been received.
        performance.mark("done");
        return;
      }

      const markEntries = entryList.getEntriesByType("mark");
      if (markEntries.length > 0) {
        assert_equals(img_entries, 1);
        t.done();
        return;
      }
    }
  );
  observer.observe({entryTypes: ["resource", "mark"]});

  // Images are added dynamically to make sure the observer is registered before their download finishes.
  const img1 = document.createElement("img");
  img1.src = img_url;
  document.body.appendChild(img1);
  const img2 = document.createElement("img");
  img2.src = img_url;
  document.body.appendChild(img2);
}, "Only one resource entry per resource");
</script>
<h1>One resource when reusing data</h1>
<p>
If the user agent is to reuse the data from another existing or completed fetch initiated from the current document, abort the remaining steps.
</p>
<div id="log"></div>
</body>
