LayoutTests/imported/w3c/web-platform-tests/resource-timing/single-entry-per-resource-expected.txt


One resource when reusing data

If the user agent is to reuse the data from another existing or completed fetch initiated from the current document, abort the remaining steps.


FAIL Only one resource entry per resource assert_equals: expected 1 but got 0
